// medical-clinics-frontend/src/app/services/models/create-physician.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreatePhysician {
  abbreviation?: string;
  description?: string;
  email: string;
  firstName: string;
  identificationNumber: string;
  lastName: string;
  pictureUrl?: string;

  /**
   * Case insensitive if word separator is space dot or hyphens will be change automatically
   */
  specialty: 'DERMATOLOGY' | 'FAMILY_MEDICINE' | 'INTERNAL_MEDICINE' | 'GYNAECOLOGY' | 'NEUROLOGY' | 'ALLERGIST' | 'CARDIOLOGY' | 'PSYCHIATRY' | 'GASTROENTEROLOGY' | 'PEDIATRICS' | 'ENDOCRINOLOGY' | 'OPHTHALMOLOGY' | 'ORTHOPEDICS';
  workplaceAddress: string;
  workplaceCity: string;
}